BREAK apt: cache: abandon hope and use Florist to track apt update calls
Running "apt-get update" can take more than 20s, which can be more than the total time
taken by Florist. For this reason, we avoid running an update if the cache age is less
than a specified validity.

Function apt.isCacheValid went through many implementations, attempting to look at the
last modified time of many files generated or updated one time or another by the APT
machinery. No matter what we did, we always encountered a corner case, or something that
worked on Ubuntu did not work on Debian, or did not work on AWS.

For this reason, we now use a logic that is guaranteed to be 100% correct and
ironically is also simpler, although it will not detect if the APT cache has been
updated OOB of Florist.