BREAK provisioner: keep going during configure, collect errors
The configure subcommand invokes in sequence:

- preConfigure
- for each flower: flower.Init, flower.Configure
- postConfigure

Before this change, it was terminating the execution on the first error encountered.

With this change, configure will invoke the same sequence, but collect the errors
instead of terminating. The errors will also be available to the postConfigure callback
via the new Provider.Errors method.

On return from postConfigure, configure will consider the collected errors to decide
which exit status code to use.

In particular, this will allow the user-provided postConfigure callback to be always
executed, enabling user-specific recovery.