Taskfile: add mutagen targets for pulumi, integrate with pulumi:up, bring coverage back
And run tests by compiling them on the target, no need to build on workstation
and then upload. This is not only simpler, it also gives a better output.

The coverage data in florist/bin/coverage.out is now synchronized back to the workstation,
so it is possible to do:

    task pulumi:test coverage:browser

Target 'pulumi:test' will run the tests on the target, mutagen will immediately synch
file florist/bin/coverage.out to the workstation, and target 'coverage:browser' will
open a web browser on the workstation showing the coverage from the target :-)